refactor!: extract generator core to @woodland-generators/core
Move src/{character,details,generators,logging,maths} and matching
test trees into packages/core/, with its own package.json, tsconfig,
and Jest configs. The package is consumed via workspace symlink: the
root CLI now imports from @woodland-generators/core instead of
relative paths. ESLint, codecov, and the bench/CI workflows are
updated to point at the new locations.

BREAKING CHANGE: the woodland-generators npm package no longer exposes
a library API at the root (main/types fields removed). The CLI binary
is unaffected. Library consumers should depend on
@woodland-generators/core directly.
