+++ .github/workflows/ci.yml
@@ -45,7 +45,7 @@ jobs:
         if: matrix.coverage
         uses: codecov/codecov-action@57e3a136b779b570ffcdbf80b3bdc90e7fab3de2 # v6
         with:
-          files: ./coverage/lcov.info
+          files: ./packages/core/coverage/lcov.info
           flags: unittests
           name: codecov-umbrella
           fail_ci_if_error: true
@@ -56,7 +56,7 @@ jobs:
         if: always() && matrix.coverage
         uses: codecov/test-results-action@0fa95f0e1eeaafde2c782583b36b28ad0d8c77d3 # v1
         with:
-          files: ./test-results/junit.xml
+          files: ./packages/core/test-results/junit.xml
           flags: unittests
           name: test-results
         env:
@@ -83,5 +83,7 @@ jobs:
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7
         with:
           name: dist
-          path: dist/
+          path: |
+            dist/
+            packages/*/dist/
           retention-days: 7
