+++ package.json
@@ -2,8 +2,6 @@
   "name": "woodland-generators",
   "version": "0.2.0",
   "description": "CLI tool for generating resources for Root: The Tabletop RPG",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
   "bin": {
     "woodland-gen": "dist/cli.js"
   },
@@ -33,39 +31,31 @@
   },
   "packageManager": "pnpm@10.33.2",
   "dependencies": {
+    "@woodland-generators/core": "workspace:*",
     "commander": "^14.0.0",
-    "just-percentile": "^4.2.0",
-    "pino": "^9.7.0",
-    "pino-pretty": "^13.1.1",
-    "seeded-random-utilities": "^1.1.4"
+    "pino": "^9.7.0"
   },
   "devDependencies": {
     "@eslint/js": "10.0.1",
-    "@types/jest": "29.5.14",
     "@types/node": "24.12.2",
     "@typescript-eslint/eslint-plugin": "8.59.2",
     "@typescript-eslint/parser": "8.59.2",
     "depcheck": "1.4.7",
     "eslint": "10.3.0",
-    "fast-check": "4.7.0",
-    "jest": "29.7.0",
-    "jest-bench": "29.7.1",
-    "jest-junit": "17.0.0",
     "npm-check-updates": "22.1.0",
     "prettier": "3.8.3",
-    "ts-jest": "29.4.9",
     "tsx": "4.21.0",
     "typescript": "5.9.3"
   },
   "scripts": {
-    "build": "tsc",
+    "build": "pnpm -r build && tsc",
     "build:watch": "tsc --watch",
     "cli": "pnpm run build && node dist/cli.js",
-    "clean": "rm -rf dist/",
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
-    "test:watch": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watch",
-    "test:coverage": "node --experimental-vm-modules node_modules/jest/bin/jest.js --coverage",
-    "bench": "node --experimental-vm-modules node_modules/jest/bin/jest.js --config jest.benchmark.config.json && pnpm exec tsx scripts/process-benchmarks.ts",
+    "clean": "rm -rf dist/ packages/*/dist/",
+    "test": "pnpm -r test",
+    "test:watch": "pnpm -r test:watch",
+    "test:coverage": "pnpm -r test:coverage",
+    "bench": "pnpm -r bench && pnpm exec tsx scripts/process-benchmarks.ts",
     "prepare": "pnpm run build",
     "prepublishOnly": "pre-commit run --all-files",
     "prepack": "pnpm run clean && pnpm run build",
