+++ packages/core/jest.config.json
@@ -2,7 +2,7 @@
   "preset": "ts-jest",
   "testEnvironment": "node",
   "clearMocks": true,
-  "collectCoverageFrom": ["src/**/*.ts", "!src/**/*.d.ts", "!src/cli.ts"],
+  "collectCoverageFrom": ["src/**/*.ts", "!src/**/*.d.ts"],
   "testPathIgnorePatterns": ["/node_modules/", "/dist/", ".*\\.bench\\.ts$"],
   "coverageReporters": ["text", "lcov", "html"],
   "reporters": [
