+++ packages/core/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "@woodland-generators/core",
+  "version": "0.0.0",
+  "description": "Reusable generator algorithms for Root: The Tabletop RPG (engine, no UI)",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "author": "Alex Brandt",
+  "license": "MIT",
+  "private": true,
+  "files": [
+    "dist/",
+    "README.md",
+    "LICENSE"
+  ],
+  "engines": {
+    "node": ">=20.0.0",
+    "pnpm": ">=10.0.0"
+  },
+  "dependencies": {
+    "just-percentile": "^4.2.0",
+    "pino": "^9.7.0",
+    "pino-pretty": "^13.1.1",
+    "seeded-random-utilities": "^1.1.4"
+  },
+  "devDependencies": {
+    "@types/jest": "29.5.14",
+    "@types/node": "24.12.2",
+    "fast-check": "4.7.0",
+    "jest": "29.7.0",
+    "jest-bench": "29.7.1",
+    "jest-junit": "17.0.0",
+    "ts-jest": "29.4.9",
+    "typescript": "5.9.3"
+  },
+  "scripts": {
+    "build": "tsc",
+    "build:watch": "tsc --watch",
+    "clean": "rm -rf dist/",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "test:watch": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watch",
+    "test:coverage": "node --experimental-vm-modules node_modules/jest/bin/jest.js --coverage",
+    "bench": "node --experimental-vm-modules node_modules/jest/bin/jest.js --config jest.benchmark.config.json"
+  }
+}
