+++ packages/core/src/character/index.ts
