+++ packages/core/src/character/types.ts
