+++ packages/core/src/details/index.ts
