+++ packages/core/src/details/types.ts
