+++ packages/core/src/generators/character.ts
