+++ packages/core/src/generators/core/choices.ts
