+++ packages/core/src/generators/demeanor.ts
