+++ packages/core/src/generators/details.ts
