+++ packages/core/src/logging/pino-pretty-transport.js
