+++ packages/core/src/maths/stats.ts
