+++ packages/core/test/generators/background.test.ts
