+++ packages/core/test/generators/character.test.ts
