+++ packages/core/test/generators/core/choices.properties.test.ts
