+++ packages/core/test/generators/core/choices.test.ts
