+++ packages/core/test/generators/move.test.ts
