+++ packages/core/test/generators/name.test.ts
