+++ packages/core/test/generators/species.properties.test.ts
