+++ packages/core/tsconfig.json
@@ -23,7 +23,7 @@
     "skipLibCheck": true,
     "removeComments": true,
     "newLine": "lf",
-    "types": ["node"]
+    "types": ["node", "jest"]
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist"]
