--- src/index.ts
@@ -1,2 +1,0 @@
-export * from "./character";
-export * from "./generators";
